requirements.txt
# 核心依赖
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.10

# 数据库和缓存
redis>=5.0.0
APScheduler>=3.10.0

# 命令行工具
rich>=13.7.0
click>=8.1.7
colorama>=0.4.6
prompt-toolkit>=3.0.43
tqdm>=4.66.0

# API 客户端
openai>=1.12.0
deepl>=1.16.0
praw>=7.7.0

# 自然语言处理
spacy>=3.7.0
textblob>=0.17.0
scikit-learn>=1.4.0

# 数据处理
pandas>=2.2.0
pydantic>=2.6.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# 网页抓取增强（可选）
selenium>=4.17.0
playwright>=1.41.0
